Restructure project file hierarchy
Place everything related to a single analyzer/test in their own
contained source file.
Use a file structure that makes more sense for a Python project.
