+++ abi-extract-info/analyzers/bitfield.py
@@ -1,9 +1,12 @@
-#! /bin/env python
 # Copyright 2025-present, Synopsys, Inc.
 # All rights reserved.
 #
 # This source code is licensed under the GPL-3.0 license found in
 # the LICENSE file in the root directory of this source tree.
+
+import random
+
+import helper
 
 """
 The purpose of this class is to create a bit-field test case
@@ -61,10 +64,6 @@ For short example:
 
     The memory layout is `0x0F` `0xFF` `0x03` `0xFF`
 """
-
-from lib import helper
-import random
-
 class BitFieldGenerator:
     def __init__(self, Target):
         self.Target = Target
@@ -525,8 +524,133 @@ class BitFieldGenerator:
         self.generate_main()
         return self.get_result()
 
-def generate(Target):
-    return BitFieldGenerator(Target).generate()
+"""
+This class parses the bit-field test case to a summary.
+It validates how bit-fields are mapped in memory.
+"""
+class BitFieldTests:
+    def __init__(self):
+        self.results = []
 
-if __name__ == "__main__":
-    print(generate())
+    def append(self, W):
+        self.results.append(W)
+
+    def get_results(self):
+        return "\n".join(self.results)
+
+    # e.g
+    # [{'dtype': ['short'], 'sign': '>', 'content': ['', '',
+    #                                                'No extra padding.:Little-endian.Extra padding.:Little-endian.']},
+    #  {'dtype': ['short'], 'sign': '<', 'content': ['No extra padding.:Little-endian.',
+    #                                                'No extra padding.:Little-endian.',
+    #                                                'No extra padding.:Little-endian.']},
+    #  {'dtype': ['int'], 'sign': '>', 'content': ['Extra padding.:Little-endian.',
+    #                                              'Extra padding.:Little-endian.',
+    #                                              'Extra padding.:Little-endian.']},
+    #  {'dtype': ['int'], 'sign': '<', 'content': ['No extra padding.:Little-endian.',
+    #                                              'No extra padding.:Little-endian.',
+    #                                              'No extra padding.:Little-endian.']},
+    #  ...
+    def process_stage1(self, content):
+        # Split the input content into individual entries
+        data = content.split("\n")
+
+        entries = []
+        for entry in data:
+            if not entry:
+                continue
+
+            # Split the string to get dtype, sign, and content
+            sdata   = entry.split(":")
+            dtype   = sdata[0][:-2]  # Removing the '_0', '_1', etc. from dtype
+            sign    = sdata[1]
+            content = ":".join(sdata[2:])
+
+            found = False
+            for e in entries:
+                if e["dtype"][0] == dtype and e["sign"] == sign:
+                    e["content"].append(content)  # Append content to existing entry
+                    found = True
+                    break
+
+            if not found:
+                entries.append({ "dtype": [dtype], "sign": sign, "content": [content] })
+
+        return entries
+
+    # e.g
+    # [{'dtype': ['short'], 'sign': '>', 'content': 'unknown'},
+    #  {'dtype': ['short'], 'sign': '<', 'content': 'No extra padding.:Little-endian.'},
+    #  {'dtype': ['int'], 'sign': '>', 'content': 'Extra padding.:Little-endian.'},
+    #  {'dtype': ['int'], 'sign': '<', 'content': 'No extra padding.:Little-endian.'}
+    #  ...
+    def process_stage2(self, entries):
+        # Find the most frequent content for each entry
+        for entry in entries:
+            occurrences = {}
+
+            # Count occurrences of each content
+            for c in entry["content"]:
+                if c in occurrences:
+                    occurrences[c] += 1
+                else:
+                    occurrences[c] = 1
+
+            # Get the most frequent content (or 'unknown' if no clear winner)
+            max_entry = max(occurrences, key=occurrences.get)
+            entry["content"] = max_entry or "unknown"
+
+        return entries
+
+    # e.g
+    # [{'dtype': ['short'], 'sign': '>', 'content': 'unknown'},
+    #  {'dtype': ['short', 'int', 'long', 'long_long'], 'sign': '<', 'content': 'No extra padding.:Little-endian.'},
+    #  {'dtype': ['int', 'long', 'long_long'], 'sign': '>', 'content': 'Extra padding.:Little-endian.'}]
+    def process_stage3(self, entries):
+        # Merge entries based on sign and content
+        for entry in entries:
+            for entry2 in entries.copy():
+                if entry["dtype"] == entry2["dtype"]:
+                    continue
+
+                # If sign and content are the same, merge dtype
+                if entry["sign"] == entry2["sign"] and entry["content"] == entry2["content"]:
+                    entry["dtype"].extend(entry2["dtype"])
+                    entries.remove(entry2)  # Remove merged entry
+
+        return entries
+
+    def summary_results(self, entries):
+        self.append("Bit-Field test:")
+        for sign in [">", "<"]:
+            self.append(f"- sum(bit-fields) {sign} sizeof(datatype)")
+            for entry in entries:
+                if entry["sign"] == sign:
+                    self.append(f"  - {' : '.join(entry['dtype'])}")
+                    tmp = entry["content"].replace(":", "\n    - ")
+                    self.append(f"    - {tmp}")
+
+    def prepare_summary(self, content):
+        entries = self.process_stage1(content)
+        entries = self.process_stage2(entries)
+        entries = self.process_stage3(entries)
+        self.summary_results(entries)
+        return self.get_results()
+
+def do_bitfield(Driver, Report, Target):
+    content = BitFieldGenerator(Target).generate()
+    open("tmp/out_bitfield.c", "w").write(content)
+    source_files   = ["tmp/out_bitfield.c"]
+    assembly_files = []
+    output_name = "out_bitfield"
+    res, stdout_file = Driver.run(source_files, assembly_files, output_name)
+    if res != 0:
+        print("Skip: Bitfields test case failed.")
+        return
+
+    content = helper.read_file(stdout_file)
+
+    summary_content = BitFieldTests().prepare_summary(content)
+    summary_file = "tmp/out_bitfield.sum"
+    open(summary_file, "w").write(summary_content)
+    Report.append(summary_file)
