+++ abi-extract-info/analyzers/datatypes.py
@@ -1,9 +1,12 @@
-#! /bin/env python
 # Copyright 2025-present, Synopsys, Inc.
 # All rights reserved.
 #
 # This source code is licensed under the GPL-3.0 license found in
 # the LICENSE file in the root directory of this source tree.
+
+import re
+
+import helper
 
 Types = [ "char", "signed char", "unsigned char", "short", "int", "long", "long long",
           "void*", "float", "double", "long double",
@@ -16,7 +19,7 @@ class DataTypesGenerator:
 
     def append(self, W):
         self.Result.append(W)
-        
+
     def generateBase(self):
         self.append("""
 #include <stdio.h>
@@ -100,6 +103,94 @@ int main() {
         self.generateMain()
 
         return self.getResult()
-    
-def generate():
-    return DataTypesGenerator().generate()
+
+class DatatypesTests:
+    def __init__(self):
+        self.result = []
+        self.categorizes = {
+            "size": {},
+            "align": {},
+            "signedness": [],
+            "struct size": {},
+            "struct align": {},
+            "union size": {},
+            "union align": {}
+        }
+
+
+    def append(self, W):
+        self.result.append(W)
+
+    def get_result(self):
+        return "\n".join(self.result)
+
+    def pattern(self):
+        return r'(\w[\w*\s]+)+:\s+signedness:\s+(\d),\s+size:\s+(\d+),\s+align:\s+(\d+)'
+
+    def find_matches(self, content):
+        return re.findall(self.pattern(), content)
+
+    def categorize_data(self, matches):
+        for data_type, signedness, size, alignment in matches:
+            data_type = data_type.strip()
+            size, alignment = str(size), str(alignment)
+
+            if "struct" in data_type:
+                base_type  = data_type[7:]
+                key_prefix = "struct"
+            elif "union" in data_type:
+                base_type = data_type[6:]
+                key_prefix = "union"
+            else:
+                base_type = data_type
+                key_prefix = ""
+
+            # Update relevant categorizes
+            if key_prefix:
+                self.categorizes[f"{key_prefix} size"].setdefault(size, []).append(base_type)
+                self.categorizes[f"{key_prefix} align"].setdefault(alignment, []).append(base_type)
+            else:
+                self.categorizes["size"].setdefault(size, []).append(base_type)
+                self.categorizes["align"].setdefault(alignment, []).append(base_type)
+                if signedness == "1":
+                    self.categorizes["signedness"].append(base_type)
+
+    # Generate summary.
+    def generate_summary(self):
+        for key in ["size", "align", "signedness", "struct size", "struct align", "union size", "union align"]:
+            self.append(f"Datatype {key} test:")
+            if key == "signedness":
+                self.append(f" - {' : '.join(self.categorizes[key])}")
+            else:
+                for size, types in self.categorizes[key].items():
+                    self.append(f" - {size}: {' : '.join(types)}")
+            self.append("")
+
+    def generate(self, content):
+        matches = self.find_matches(content)
+        self.categorize_data(matches)
+        self.generate_summary()
+
+        return self.get_result()
+
+def do_datatypes(Driver, Report, Target):
+    Content = DataTypesGenerator().generate()
+    open("tmp/out_datatypes.c", "w").write(Content)
+    source_files = ["tmp/out_datatypes.c"]
+    assembly_files = []
+    output_name = "out_datatypes"
+    res, stdoutFile = Driver.run(source_files, assembly_files, output_name)
+    if res != 0:
+        print("fatal: Datatypes test case failed. Enable verbose mode (-v) for more details.")
+        exit(1)
+
+    # Reading the file generated by the C program executed.
+    Stdout = helper.read_file(stdoutFile)
+    # Parsing the type details (signedness/size/align) into the 'targetArch' class.
+    Target.set_type_details(helper.parse_type_info(Stdout))
+
+    content = DatatypesTests().generate(Stdout)
+    open("tmp/out_datatypes.sum", "w").write(content)
+
+    # Store the generated report file for datatypes test case.
+    Report.append("tmp/out_datatypes.sum")
