+++ abi-extract-info/analyzers/saved.py
@@ -1,11 +1,12 @@
-#!/bin/env python
 # Copyright 2025-present, Synopsys, Inc.
 # All rights reserved.
 #
 # This source code is licensed under the GPL-3.0 license found in
 # the LICENSE file in the root directory of this source tree.
 
-from lib import helper
+import helper
+import hexUtils
+import dumpInformation
 
 """
 The purpose of this generator is to create a test case to
@@ -94,12 +95,66 @@ int main (void) {
 
         return self.get_result()
 
-def generate_main(Target, hvalue_caller_saved):
-    return ReturnGenerator(Target).generate_main(hvalue_caller_saved)
+class SavedTests:
+    def __init__(self, Target):
+        self.Target = Target
 
-def generate_aux(Target, hvalue_callee_saved):
-    return ReturnGenerator(Target).generate_aux(hvalue_callee_saved)
+    def generate_summary(self, caller, callee):
 
-if __name__ == "__main__":
-    content = generate(None, None, 'int', 4)
-    print(content)
+        caller_str = ", ".join(caller)
+        callee_str = ", ".join(callee)
+
+        summary = []
+        summary.append("Caller/callee-saved test:")
+        summary.append(f" - caller-saved {caller_str}")
+        summary.append(f" - callee-saved {callee_str}")
+        summary.append("")
+
+        return "\n".join(summary)
+
+    def run_test(self, register_banks, hvalue_caller_saved, hvalue_callee_saved):
+        hutils = hexUtils.HexUtils(self.Target)
+
+        tmp = hutils.find_registers_fill([hvalue_caller_saved], register_banks)
+        caller_saved_registers, _  = tmp
+
+        tmp = hutils.find_registers_fill([hvalue_callee_saved], register_banks)
+        callee_saved_registers, _ = tmp
+
+        return self.generate_summary(caller_saved_registers, callee_saved_registers)
+def do_saved(Driver, Report, Target):
+    sizeof = Target.get_type_details("int")["size"]
+
+    # Generate value for callee/caller saved registers.
+    helper.reset_used_values()
+    hvalue_caller_saved = helper.generate_hexa_value(sizeof)
+    hvalue_callee_saved = helper.generate_hexa_value(sizeof)
+
+    content = ReturnGenerator(Target).generate_main(hvalue_caller_saved)
+    open("tmp/out_saved_main.c", "w").write(content)
+    content = ReturnGenerator(Target).generate_aux(hvalue_callee_saved)
+    open("tmp/out_saved_aux.c", "w").write(content)
+
+    source_files = ["tmp/out_saved_main.c", \
+                    "tmp/out_saved_aux.c", "src/helper.c"]
+    assembly_files = ["src/arch/riscv.S"]
+    output_name = "out_saved"
+    res, stdout_file = Driver.run(source_files, assembly_files, output_name)
+    if res != 0:
+        print("Skip: Caller/Callee-saved test case failed.")
+        return
+
+    dump_information = dumpInformation.DumpInformation()
+    dump_information.parse(stdout_file, True)
+    register_banks = dump_information.get_reg_banks()
+
+    saved_tests = SavedTests(Target)
+    summary_content = saved_tests.run_test(register_banks, \
+                                hvalue_caller_saved, hvalue_callee_saved)
+
+    summary_file = "tmp/out_saved.sum"
+    open(summary_file, "w").write(summary_content)
+
+    # Store the generated report file for argument passing test case.
+    Report.append(summary_file)
+
