+++ abi-extract-info/dumpInformation.py
