+++ scripts/wrapper/cc/clang-rv32gc-ilp32/cc-wrapper
