+++ scripts/wrapper/cc/clang-rv32gc-ilp32d/ld-wrapper
