+++ scripts/wrapper/cc/gcc-rv32gc-ilp32/cc-wrapper
