+++ scripts/wrapper/cc/gcc-rv32gc-ilp32/ld-wrapper
