+++ scripts/wrapper/cc/gcc-rv32gc-ilp32d/cc-wrapper
