+++ scripts/wrapper/cc/gcc-rv32gc-ilp32d/ld-wrapper
