+++ src/stack_dir/A.c
