+++ src/stack_dir/main.c
